The function shuffledImagesMemoBlocks is created
